feat(ios): modular header import support
