+++ framework/examples/ios-demo/HippyDemo/turbomodule/TurboConfig.m
@@ -22,7 +22,6 @@
 
 #import "TurboConfig.h"
 #import "HippyDefines.h"
-#import "HippyDefines.h"
 
 @interface TurboConfig ()
